+++ src/data/image/intensity/siff/unregistered.rs
@@ -65,7 +65,7 @@ pub fn load_array_raw_siff<T : Into<u64>>(
 
 #[binrw::parser(reader)]
 pub fn extract_mask_raw_siff<T : Into<u64>>(
-    target_array : &mut ArrayViewMut1<u64>,
+    target_array : &mut ArrayViewMut1<u16>,
     mask : &ArrayView2<bool>,
     lookup_table : &ArrayView2<usize>,
     strip_bytes : T,
@@ -80,7 +80,7 @@ pub fn extract_mask_raw_siff<T : Into<u64>>(
                 photon_to_y!(siffphoton, 0, ydim),
                 photon_to_x!(siffphoton, 0, xdim)
             );
-            target_array[lookup_table[[y, x]]] += mask[[y, x]] as u64;
+            target_array[lookup_table[[y, x]]] += mask[[y, x]] as u16;
         }
     );
     Ok(())
@@ -171,7 +171,7 @@ pub fn load_array_compressed_siff(
 
 #[binrw::parser(reader, endian)]
 pub fn extract_mask_compressed_siff(
-    target_array : &mut ArrayViewMut1<u64>,
+    target_array : &mut ArrayViewMut1<u16>,
     mask : &ArrayView2<bool>,
     lookup_table : &ArrayView2<usize>,
     ydim : u32,
@@ -188,7 +188,7 @@ pub fn extract_mask_compressed_siff(
     for (&mask_px, &frame_px, &lookup_px) in izip!(
         mask.iter(), frame_array.iter(), lookup_table.iter()
     ) {
-       target_array[lookup_px] += mask_px as u64 * frame_px as u64;
+       target_array[lookup_px] += mask_px as u16 * frame_px as u16;
     }
 
     Ok(())
